rc-str 前に低依存プリミティブをカテゴリ別モジュールに分割する
`src/primitives.rs` のうち、Plan D / rc-str と衝突しにくい低依存の単純プリミティブを
`src/primitives/` 配下のカテゴリ別モジュール (stack / numeric / logic) に切り出す。

- `src/primitives/stack.rs`: drop_prim, lit_marker_prim, dup_prim, swap_prim
- `src/primitives/numeric.rs`: add_prim, sub_prim, mul_prim, div_prim, mod_prim,
  sqrt_prim, negate_prim
- `src/primitives/logic.rs`: and_prim, or_prim, band_prim, bor_prim

`src/primitives.rs` はファサードとして残し、`mod ...;` と `pub use ...::*;` で
既存の `crate::primitives::<name>` 参照を維持する。compare / memory / strings /
arrays / register_all 周辺は触らない。

挙動変更なし。
